Refactoring subcommands
