+++ genai-perf/genai_perf/export_data/output_reporter.py
@@ -30,7 +30,7 @@ from typing import Optional
 from genai_perf.export_data.data_exporter_factory import DataExporterFactory
 from genai_perf.export_data.exporter_config import ExporterConfig
 from genai_perf.metrics import Statistics, TelemetryStatistics
-from genai_perf.parser import get_extra_inputs_as_dict
+from genai_perf.subcommand.common import get_extra_inputs_as_dict
 
 
 class OutputReporter:
