+++ genai-perf/genai_perf/main.py
@@ -28,165 +28,9 @@
 import os
 import sys
 import traceback
-from argparse import Namespace
-from pathlib import Path
-from typing import Optional, Tuple
 
 import genai_perf.logging as logging
 from genai_perf import parser
-from genai_perf.checkpoint.checkpoint import Checkpoint
-from genai_perf.config.generate.genai_perf_config import GenAIPerfConfig
-from genai_perf.config.generate.perf_analyzer_config import PerfAnalyzerConfig
-from genai_perf.config.generate.search_parameters import SearchParameters
-from genai_perf.config.generate.sweep_objective_generator import SweepObjectiveGenerator
-from genai_perf.config.input.config_command import ConfigCommand, Range, Subcommand
-from genai_perf.config.run.run_config import RunConfig
-from genai_perf.constants import DEFAULT_TRITON_METRICS_URL
-from genai_perf.exceptions import GenAIPerfException
-from genai_perf.export_data.output_reporter import OutputReporter
-from genai_perf.inputs.input_constants import DEFAULT_STARTING_INDEX
-from genai_perf.inputs.inputs import Inputs
-from genai_perf.inputs.inputs_config import InputsConfig
-from genai_perf.measurements.run_config_measurement import RunConfigMeasurement
-from genai_perf.plots.plot_config_parser import PlotConfigParser
-from genai_perf.plots.plot_manager import PlotManager
-from genai_perf.profile_data_parser import (
-    ImageRetrievalProfileDataParser,
-    LLMProfileDataParser,
-    ProfileDataParser,
-)
-from genai_perf.telemetry_data.triton_telemetry_data_collector import (
-    TelemetryDataCollector,
-    TritonTelemetryDataCollector,
-)
-from genai_perf.tokenizer import Tokenizer, get_tokenizer
-from genai_perf.types import GpuRecords, ModelObjectiveParameters
-
-
-def create_artifacts_dirs(args: Namespace) -> None:
-    plot_dir = args.artifact_dir / "plots"
-    os.makedirs(args.artifact_dir, exist_ok=True)
-    if hasattr(args, "generate_plots") and args.generate_plots:
-        os.makedirs(plot_dir, exist_ok=True)
-
-
-def create_config_options(args: Namespace) -> InputsConfig:
-
-    try:
-        extra_input_dict = parser.get_extra_inputs_as_dict(args)
-    except ValueError as e:
-        raise GenAIPerfException(e)
-
-    return InputsConfig(
-        input_type=args.prompt_source,
-        output_format=args.output_format,
-        model_name=args.model,
-        model_selection_strategy=args.model_selection_strategy,
-        input_filename=args.input_file,
-        synthetic_input_filenames=args.synthetic_input_files,
-        starting_index=DEFAULT_STARTING_INDEX,
-        length=args.num_prompts,
-        prompt_tokens_mean=args.synthetic_input_tokens_mean,
-        prompt_tokens_stddev=args.synthetic_input_tokens_stddev,
-        output_tokens_mean=args.output_tokens_mean,
-        output_tokens_stddev=args.output_tokens_stddev,
-        output_tokens_deterministic=args.output_tokens_mean_deterministic,
-        image_width_mean=args.image_width_mean,
-        image_width_stddev=args.image_width_stddev,
-        image_height_mean=args.image_height_mean,
-        image_height_stddev=args.image_height_stddev,
-        image_format=args.image_format,
-        random_seed=args.random_seed,
-        num_prompts=args.num_prompts,
-        add_stream=args.streaming,
-        tokenizer=get_tokenizer(
-            args.tokenizer, args.tokenizer_trust_remote_code, args.tokenizer_revision
-        ),
-        extra_inputs=extra_input_dict,
-        batch_size_image=args.batch_size_image,
-        batch_size_text=args.batch_size_text,
-        output_dir=args.artifact_dir,
-    )
-
-
-def create_telemetry_data_collector(
-    args: Namespace,
-) -> Optional[TelemetryDataCollector]:
-    logger = logging.getLogger(__name__)
-    telemetry_data_collector = None
-    if args.service_kind == "triton":
-        server_metrics_url = args.server_metrics_url or DEFAULT_TRITON_METRICS_URL
-        telemetry_data_collector = TritonTelemetryDataCollector(server_metrics_url)
-
-    if telemetry_data_collector and not telemetry_data_collector.is_url_reachable():
-        logger.warning(
-            f"The metrics URL ({telemetry_data_collector.metrics_url}) is unreachable. "
-            "GenAI-Perf cannot collect telemetry data."
-        )
-        telemetry_data_collector = None
-    return telemetry_data_collector
-
-
-def generate_inputs(config_options: InputsConfig) -> None:
-    inputs = Inputs(config_options)
-    inputs.create_inputs()
-
-
-def calculate_metrics(args: Namespace, tokenizer: Tokenizer) -> ProfileDataParser:
-    if args.endpoint_type in ["embeddings", "nvclip", "rankings"]:
-        return ProfileDataParser(
-            args.profile_export_file,
-            goodput_constraints=args.goodput,
-        )
-    elif args.endpoint_type == "image_retrieval":
-        return ImageRetrievalProfileDataParser(
-            args.profile_export_file,
-            goodput_constraints=args.goodput,
-        )
-    else:
-        return LLMProfileDataParser(
-            filename=args.profile_export_file,
-            tokenizer=tokenizer,
-            goodput_constraints=args.goodput,
-        )
-
-
-def report_output(
-    data_parser: ProfileDataParser,
-    telemetry_data_collector: Optional[TelemetryDataCollector],
-    args: Namespace,
-) -> None:
-    if args.concurrency:
-        infer_mode = "concurrency"
-        load_level = f"{args.concurrency}"
-    elif args.request_rate:
-        infer_mode = "request_rate"
-        load_level = f"{args.request_rate}"
-    else:
-        raise GenAIPerfException("No valid infer mode specified")
-
-    stats = data_parser.get_statistics(infer_mode, load_level)
-    telemetry_stats = (
-        telemetry_data_collector.get_statistics() if telemetry_data_collector else None
-    )
-    reporter = OutputReporter(stats, telemetry_stats, args)
-
-    reporter.report_output()
-    if args.generate_plots:
-        create_plots(args)
-
-
-def create_plots(args: Namespace) -> None:
-    # TMA-1911: support plots CLI option
-    plot_dir = args.artifact_dir / "plots"
-    PlotConfigParser.create_init_yaml_config(
-        filenames=[args.profile_export_file],  # single run
-        output_dir=plot_dir,
-    )
-    config_parser = PlotConfigParser(plot_dir / "config.yaml")
-    plot_configs = config_parser.generate_configs(args.tokenizer)
-    plot_manager = PlotManager(plot_configs)
-    plot_manager.generate_plots()
 
 
 # Separate function that can raise exceptions used for testing
@@ -198,152 +42,9 @@ def run():
     if args.subcommand == "compare":
         args.func(args)
     elif args.subcommand == "analyze":
-        analyze_subcommand(args)
-    else:
-        config_options = create_config_options(args)
-        create_artifacts_dirs(args)
-        tokenizer = get_tokenizer(
-            args.tokenizer,
-            args.tokenizer_trust_remote_code,
-            args.tokenizer_revision,
-        )
-        generate_inputs(config_options)
-        telemetry_data_collector = create_telemetry_data_collector(args)
-        args.func(args, extra_args, telemetry_data_collector)
-        data_parser = calculate_metrics(args, tokenizer)
-        report_output(data_parser, telemetry_data_collector, args)
-
-
-def analyze_subcommand(args: Namespace) -> None:
-    #
-    # Setup
-    config = _setup_config(args)
-
-    model_name = config.model_names[0]
-    model_search_parameters = {
-        model_name: SearchParameters(config=config, subcommand=Subcommand.ANALYZE)
-    }
-    sweep_objective_generator = SweepObjectiveGenerator(config, model_search_parameters)
-    telemetry_data_collector = create_telemetry_data_collector(args)
-    checkpoint = Checkpoint(config)
-    results = checkpoint.results
-
-    #
-    # Sweep Loop
-    for count, objectives in enumerate(sweep_objective_generator.get_objectives()):
-        #
-        # Create GAP/PA Configs
-        genai_perf_config = GenAIPerfConfig(
-            config=config, args=args, model_objective_parameters=objectives
-        )
-        # The GAP/PA Configs will (for now) modify the CLI args
-        # based on the objective being swept
-        gap_obj_args = genai_perf_config.get_obj_args()
-
-        perf_analyzer_config = PerfAnalyzerConfig(
-            model_name=model_name,
-            args=gap_obj_args,
-            config=config,
-            model_objective_parameters=objectives,
-        )
-        obj_args = perf_analyzer_config.get_obj_args()
-
-        #
-        # Create Input/Artifacts
-        input_config_options = create_config_options(obj_args)
-        create_artifacts_dirs(obj_args)
-        tokenizer = get_tokenizer(
-            obj_args.tokenizer,
-            obj_args.tokenizer_trust_remote_code,
-            obj_args.tokenizer_revision,
-        )
-        generate_inputs(input_config_options)
-
-        #
-        # Run PA
-        args.func(obj_args, perf_analyzer_config, telemetry_data_collector)
-
-        #
-        # Extract Perf Metrics
-        infer_mode, load_level = _determine_infer_mode_and_load_level(
-            obj_args, objectives, model_name
-        )
-        data_parser = calculate_metrics(obj_args, tokenizer)
-        perf_stats = data_parser.get_statistics(infer_mode, load_level)
-        perf_metrics = perf_stats.create_records()
-
-        #
-        # Extract Telemetry Metrics
-        # FIXME: Once I'm able to collect telemetry records will need
-        # to write a method to hook this up
-        # telemetry_stats = (
-        #     telemetry_data_collector.get_statistics()
-        #     if telemetry_data_collector
-        #     else None
-        # )
-        gpu_metrics: GpuRecords = {}
-
-        #
-        # Create RunConfigMeasurement
-        run_config_measurement = RunConfigMeasurement(gpu_metrics)
-        run_config_measurement.add_perf_metrics(model_name, perf_metrics)
-
-        #
-        # Create RunConfig
-        run_config_name = model_name + "_run_config_" + str(count)
-        run_config = RunConfig(
-            name=run_config_name,
-            genai_perf_config=genai_perf_config,
-            perf_analyzer_config=perf_analyzer_config,
-            measurement=run_config_measurement,
-        )
-
-        #
-        # Add to results and write checkpoint
-        results.add_run_config(run_config)
-        checkpoint.create_checkpoint_object()
-
-
-def _setup_config(args: Namespace) -> ConfigCommand:
-    config = ConfigCommand(model_names=args.model)
-
-    if args.sweep_list:
-        config.analyze.sweep_parameters = {args.sweep_type: args.sweep_list}
-    else:
-        config.analyze.sweep_parameters = {
-            args.sweep_type: Range(min=args.sweep_min, max=args.sweep_max)
-        }
-
-    return config
-
-
-def _determine_infer_mode_and_load_level(
-    args: Namespace, objectives: ModelObjectiveParameters, model_name: str
-) -> Tuple[str, str]:
-    if args.sweep_type == "concurrency":
-        infer_mode = "concurrency"
-        load_level = (
-            f"{objectives[model_name][infer_mode].get_value_based_on_category()}"
-        )
-    elif args.sweep_type == "request_rate":
-        infer_mode = "request_rate"
-        load_level = (
-            f"{float(objectives[model_name][infer_mode].get_value_based_on_category())}"
-        )
-    elif args.sweep_type == "input_sequence_length" or args.sweep_type == "num_prompts":
-        if args.concurrency:
-            infer_mode = "concurrency"
-            load_level = f"{args.concurrency}"
-        elif args.request_rate:
-            infer_mode = "request_rate"
-            load_level = f"{args.request_rate}"
-        else:
-            infer_mode = "concurrency"
-            load_level = "1"
-    else:
-        raise GenAIPerfException("Cannot determine infer_mode/load_level")
-
-    return infer_mode, load_level
+        args.func(args)
+    else:  # profile
+        args.func(args, extra_args)
 
 
 def main():
